refactor etl job
